+++ src/vtk_support.cpp
@@ -23,8 +23,8 @@ uint8_t VTK_QUADRATIC_HEXAHEDRON = 25;
 
 // Contains data for VTK UnstructuredGrid
 struct VtkData {
-    int64_t *offset;
-    int64_t *cells;
+    int *offset;
+    int *cells;
     uint8_t *celltypes;
     int loc;   // current position within cells
     int *nref; // conversion between ansys and vtk numbering
@@ -442,8 +442,8 @@ int ans_to_vtk(
     const int *type_ref,
     const int nnode,
     const int *nnum,
-    int64_t *offset,
-    int64_t *cells,
+    int *offset,
+    int *cells,
     uint8_t *celltypes) {
     bool is_quad;
     int i;          // counter
